requirements.txt
asgiref==3.7.2
django-mssql-backend==2.8.1
django-pyodbc-azure-2019==2.1.0.0
djangorestframework==3.14.0
mssql-django==1.2
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
Django==4.0